+++ assets/js/base/components/cart-checkout/place-order-button/index.js
@@ -4,7 +4,7 @@
 import { __ } from '@wordpress/i18n';
 import { useCheckoutSubmit } from '@woocommerce/base-context/hooks';
 import { Icon, done } from '@woocommerce/icons';
-import { Button } from '@woocommerce/blocks-checkout';
+import Button from '@woocommerce/base-components/button';
 
 const PlaceOrderButton = () => {
 	const {
