+++ assets/js/base/components/cart-checkout/shipping-rates-control-package/index.tsx
@@ -6,7 +6,8 @@ import { _n, sprintf } from '@wordpress/i18n';
 import { decodeEntities } from '@wordpress/html-entities';
 import type { ReactElement } from 'react';
 import type { PackageRateOption } from '@woocommerce/type-defs/shipping';
-import { Panel, Label } from '@woocommerce/blocks-checkout';
+import { Panel } from '@woocommerce/blocks-checkout';
+import Label from '@woocommerce/base-components/label';
 import { useSelectShippingRate } from '@woocommerce/base-context/hooks';
 import type { CartShippingPackageShippingRate } from '@woocommerce/type-defs/cart';
