+++ assets/js/base/components/cart-checkout/totals/coupon/index.js
@@ -3,8 +3,9 @@
  */
 import { __ } from '@wordpress/i18n';
 import { useState, useEffect, useRef } from '@wordpress/element';
-import { Button, Panel, Label } from '@woocommerce/blocks-checkout';
+import Button from '@woocommerce/base-components/button';
 import { ValidatedTextInput } from '@woocommerce/base-components/text-input';
+import Label from '@woocommerce/base-components/label';
 import LoadingMask from '@woocommerce/base-components/loading-mask';
 import PropTypes from 'prop-types';
 import { withInstanceId } from '@woocommerce/base-hocs/with-instance-id';
@@ -12,6 +13,7 @@ import {
 	ValidationInputError,
 	useValidationContext,
 } from '@woocommerce/base-context';
+import { Panel } from '@woocommerce/blocks-checkout';
 
 /**
  * Internal dependencies
