+++ assets/js/base/components/label/index.tsx
