+++ assets/js/base/components/label/test/index.js
