+++ assets/js/base/components/text-input/text-input.tsx
@@ -4,11 +4,11 @@
 import { forwardRef, InputHTMLAttributes } from 'react';
 import classnames from 'classnames';
 import { useState } from '@wordpress/element';
-import { Label } from '@woocommerce/blocks-checkout';
 
 /**
  * Internal dependencies
  */
+import Label from '../label';
 import './style.scss';
 
 interface TextInputProps
