+++ assets/js/blocks/cart-checkout/payment-methods/payment-methods.js
@@ -3,7 +3,7 @@
  */
 import { usePaymentMethods } from '@woocommerce/base-context/hooks';
 import { __ } from '@wordpress/i18n';
-import { Label } from '@woocommerce/blocks-checkout';
+import Label from '@woocommerce/base-components/label';
 import { usePaymentMethodDataContext } from '@woocommerce/base-context';
 
 /**
