+++ packages/checkout/index.js
@@ -7,5 +7,3 @@ export { default as ExperimentalDiscountsMeta } from './discounts-meta';
 export { default as ExperimentalOrderShippingPackages } from './order-shipping-packages';
 export { default as Panel } from './panel';
 export { SlotFillProvider } from 'wordpress-components';
-export { default as Button } from './button';
-export { default as Label } from './label';
